+++ src/client/src/reducers/generationStatus/index.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux";
+
+import { statusMessage } from "./statusMessage";
+import { genStatus } from "./genStatus";
+
+export default combineReducers({
+  statusMessage,
+  genStatus
+});