+++ src/client/src/selectors/vscodeApiSelector.ts
@@ -1,0 +1,6 @@
+import { IVSCodeObject } from "../reducers/vscodeApiReducer";
+
+const getVSCodeApiSelector = (state: any): IVSCodeObject =>
+  state.vscode.vscodeObject;
+
+export { getVSCodeApiSelector };