+++ src/extension/src/constants.ts
@@ -83,7 +83,13 @@ export const CONSTANTS = {
     return `COSMOSDB_CONNSTR=${origin}/${username}\nCOSMOSDB_USER=${username}\nCOSMOSDB_PASSWORD=${password}\n`;
   },
   MAX_PROJECT_NAME_LENGTH: 50,
-  PORT: "5000"
+  PORT: "5000",
+  VSCODE_COMMAND: {
+    OPEN_FOLDER: "vscode.openFolder"
+  },
+  PLATFORM: {
+    WIN_32: "win32"
+  }
 };
 
 export enum ExtensionCommand {
@@ -101,7 +107,10 @@ export enum ExtensionCommand {
   GetCosmosAPIs = "getCosmosAPIs",
   GetUserStatus = "getUserStatus",
   ProjectPathAndNameValidation = "project-path-and-name-validation",
-  TrackPageSwitch = "track-page-switch"
+  TrackPageSwitch = "track-page-switch",
+  UpdateGenStatusMessage = "update-status-message",
+  UpdateGenStatus = "update-status",
+  OpenProjectVSCode = "open-project-vscode"
 }
 
 export enum TelemetryEventName {